+++ components/common/BioEdit.jsx
@@ -1,3 +1,4 @@
+"use client"
 import { Box, Container, TextField, Typography } from '@mui/material'
 import React from 'react'
 import Chip from '@mui/material/Chip';
