+++ components/common/SearchInEmploeeList.jsx
@@ -1,3 +1,4 @@
+"use client"
 import * as React from 'react';
 import TextField from '@mui/material/TextField';
 import Autocomplete, { createFilterOptions } from '@mui/material/Autocomplete';
